+++ packages/core/src/components/Heading.tsx
@@ -3,6 +3,7 @@ import classNames from "clsx";
 
 import { TextProps, CommonProps, SpacingProps, DisplayProps } from "../interfaces";
 import { ColorScheme, ColorWeight, TextVariant, SpacingToken } from "../types";
+import { numericSpacingStyle } from "../utils/spacingStyle";
 
 type HeadingProps<T extends ElementType> = TextProps<T> &
   CommonProps &
@@ -105,16 +106,22 @@ const Heading = <T extends ElementType = "h1">({
   const combinedStyle = {
     textAlign: align,
     textWrap: wrap,
-    padding: typeof padding === "number" ? `${padding}rem` : undefined,
-    paddingLeft: typeof paddingLeft === "number" ? `${paddingLeft}rem` : typeof paddingX === "number" ? `${paddingX}rem` : undefined,
-    paddingRight: typeof paddingRight === "number" ? `${paddingRight}rem` : typeof paddingX === "number" ? `${paddingX}rem` : undefined,
-    paddingTop: typeof paddingTop === "number" ? `${paddingTop}rem` : typeof paddingY === "number" ? `${paddingY}rem` : undefined,
-    paddingBottom: typeof paddingBottom === "number" ? `${paddingBottom}rem` : typeof paddingY === "number" ? `${paddingY}rem` : undefined,
-    margin: typeof margin === "number" ? `${margin}rem` : undefined,
-    marginLeft: typeof marginLeft === "number" ? `${marginLeft}rem` : typeof marginX === "number" ? `${marginX}rem` : undefined,
-    marginRight: typeof marginRight === "number" ? `${marginRight}rem` : typeof marginX === "number" ? `${marginX}rem` : undefined,
-    marginTop: typeof marginTop === "number" ? `${marginTop}rem` : typeof marginY === "number" ? `${marginY}rem` : undefined,
-    marginBottom: typeof marginBottom === "number" ? `${marginBottom}rem` : typeof marginY === "number" ? `${marginY}rem` : undefined,
+    ...numericSpacingStyle({
+      padding,
+      paddingX,
+      paddingY,
+      paddingTop,
+      paddingRight,
+      paddingBottom,
+      paddingLeft,
+      margin,
+      marginX,
+      marginY,
+      marginTop,
+      marginRight,
+      marginBottom,
+      marginLeft,
+    }),
     ...style,
   };